+++ src/terra-dev-site/test/loaders/example.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
+import classNames from 'classnames/bind';
+import styles from './example.scss';
+
+const cx = classNames.bind(styles);
 
 const TagComp = () => (
-  <h1>
+  <h1 className={cx('body')}>
     I am a test
   </h1>
 );