+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Unreleased
 ### Changed
 * Add wdio tests for additional themes, update travis ci config.
 
+### Removed
+* Removed toggle backgrounds tab and functionality from example loader
+
+### Added
+* Added CSS display tab and functionality to example loader
+
 6.25.0 - (July 7, 2020)
 ----------
 ### Fixed